+++ packages/button/src/styled/IconWrapper.jsx
@@ -18,6 +18,7 @@ const IconWrapper = (props) => {
     lineHeight: 0,
     fontSize: 0,
     margin: getMargin(props),
+    userSelect: 'none',
   };
   return <span style={style}>{props.children}</span>;
 };