+++ packages/editor-common/src/schema/default-schema.ts
@@ -46,5 +46,6 @@ export const defaultSchema: Schema = createSchema({
     'mentionQuery',
     'emojiQuery',
     'textColor',
+    'confluenceInlineComment',
   ]
 });