+++ packages/editor-core/src/plugins/media/media-plugin-options.ts
@@ -3,6 +3,7 @@ import { default as ProviderFactory } from '../../providerFactory';
 import { ErrorReporter } from '../../utils';
 
 export type MediaPluginOptions = {
+  slim?: boolean;
   providerFactory: ProviderFactory;
   errorReporter?: ErrorReporter;
   uploadErrorHandler?: (state: MediaState) => void;