+++ packages/editor-core/src/ui/ToolbarEmojiPicker/index.tsx
@@ -53,6 +53,9 @@ export default class ToolbarEmojiPicker extends PureComponent<Props, State> {
     if (this.pluginState) {
       this.pluginState.subscribe(this.handlePluginStateChange);
     }
+    // Keymapping must be added here at the document level as editor focus is lost
+    // when the picker opens so plugins/emojis/keymaps.ts will not register ESC
+    document.addEventListener('keydown', this.handleEscape);
   }
 
   componentDidUpdate() {
@@ -63,11 +66,18 @@ export default class ToolbarEmojiPicker extends PureComponent<Props, State> {
   }
 
   componentWillUnmount() {
+    document.removeEventListener('keydown', this.handleEscape);
     if (this.pluginState) {
       this.pluginState.unsubscribe(this.handlePluginStateChange);
     }
   }
 
+  handleEscape = (e) => {
+    // ESC key pressed
+    if (this.state.isOpen && (e.which === 27 || e.keyCode === 27)) {
+      this.close();
+    }
+  }
 
   private setPluginState(props: Props) {
     const { editorView, pluginKey } = props;
@@ -81,12 +91,15 @@ export default class ToolbarEmojiPicker extends PureComponent<Props, State> {
     if (pluginState) {
       this.pluginState = pluginState;
       pluginState.subscribe(this.handlePluginStateChange);
+      this.setState({
+        disabled: !pluginState.isEnabled()
+      });
     }
   }
 
   private handlePluginStateChange = (pluginState: EmojiState) => {
     this.setState({
-      disabled: !pluginState.enabled
+      disabled: !pluginState.isEnabled()
     });
   }