+++ packages/logo/test/unit/Logo.js
@@ -1,65 +1,14 @@
-import { shallow, mount } from 'enzyme';
+import { mount } from 'enzyme';
 import React from 'react';
-import { ThemeProvider } from 'styled-components';
-import Logo, { logoColoursThemeKey } from '../../src/components/LogoBase';
-import Size from '../../src/styled/Size';
+import DefaultLogo from '../../src/';
 
 describe('<Logo />', () => {
   describe('props', () => {
     it('should render an svg', () => {
-      const wrapper = mount(<Logo><svg /></Logo>);
-      expect(wrapper.find('svg')).toHaveLength(1);
-    });
-
-    describe('collapseTo prop', () => {
-      it('should not collapse by default', () => {
-        const wrapper = shallow(<Logo><svg /></Logo>);
-        expect(wrapper.props().collapseTo).toBe(undefined);
-      });
-
-      it('should pass collapseTo to the Size component', () => {
-        const wrapper = shallow(
-          <Logo collapseTo="type">
-            <svg />
-          </Logo>
-        );
-        expect(wrapper.find(Size).prop('collapseTo')).toBe('type');
-      });
-    });
-
-    describe('size prop', () => {
-      it('should be medium by default', () => {
-        const wrapper = shallow(<Logo><svg /></Logo>);
-        expect(wrapper.find(Size).props().size).toBe('medium');
-      });
-
-      it('should pass size to the Size component', () => {
-        const wrapper = shallow(<Logo size="large"><svg /></Logo>);
-        expect(wrapper.find(Size).prop('size')).toBe('large');
-      });
-    });
-
-    describe('color props', () => {
-      it('should be passed to child component via namespaced theme', () => {
-        const wrapper = shallow(
-          <Logo
-            textColor="red"
-            iconColor="blue"
-            iconGradientStart="purple"
-            iconGradientStop="green"
-          >
-            <svg />
-          </Logo>
-        );
-        expect(wrapper.find(ThemeProvider).prop('theme')).toEqual({
-          [logoColoursThemeKey]: {
-            textColor: 'red',
-            iconColor: 'blue',
-            iconGradientStart: 'purple',
-            iconGradientStop: 'green',
-          },
-        });
-      });
+      const wrapper = mount(<DefaultLogo />);
+      // Need to search the html string because we use dangerouslySetInnerHTML, which
+      // does not work with .find
+      expect(wrapper.html().includes('<svg')).toBe(true);
     });
   });
 });