+++ packages/button/docs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atlaskit/button
 
+## 5.4.8 (2017-10-27)
+
+* bug fix; correct formatting for user-select style ([fe9419c](https://bitbucket.org/atlassian/atlaskit/commits/fe9419c))
+## 5.4.7 (2017-10-27)
+
+* bug fix; change icon to be unselectable so button content can be copied ([e8c876a](https://bitbucket.org/atlassian/atlaskit/commits/e8c876a))
+## 5.4.6 (2017-10-27)
+
+* bug fix; updated button props typings ([c7a9c09](https://bitbucket.org/atlassian/atlaskit/commits/c7a9c09))
+
 ## 5.4.5 (2017-10-23)
 
 * bug fix; support false/null/undefined children in ButtonGroup ([4667228](https://bitbucket.org/atlassian/atlaskit/commits/4667228))